cvperfect_py/requirements.txt
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.2
rapidfuzz==3.5.2
typer==0.9.0
jinja2==3.1.2
pydantic==2.5.0
pytest==7.4.3
python-Levenshtein==0.20.9
bleach==6.1.0